pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "platform-cli"
version = "0.1.0"
description = "CLI for deterministic AI governance platform"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Platform Team" }]
dependencies = [
  "pydantic",
  "pyyaml",
  "requests",
  "rich",
  "typer",
]

[project.scripts]
platform = "platform_cli.main:app"

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/platform_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
